requirements.txt
requests
beautifulsoup4
pandas
streamlit
plotly
sqlalchemy
lxml
python-dotenv
